Refactor SummaryPoint to use double for total
- Change total field type from BigDecimal to double in SummaryPoint and related classes.
- Update methods and tests to accommodate the new total type.

+++ app/src/main/java/com/fabiankevin/app/models/SummaryPoint.java
@@ -1,9 +1,7 @@
 package com.fabiankevin.app.models;
 
-import java.math.BigDecimal;
-
 public record SummaryPoint(
         String label,
-        BigDecimal total
+        double total
 ) {
 }
