+++ app/src/main/java/com/fabiankevin/app/web/controllers/dtos/SummaryPointResponse.java
@@ -2,11 +2,9 @@ package com.fabiankevin.app.web.controllers.dtos;
 
 import com.fabiankevin.app.models.SummaryPoint;
 
-import java.math.BigDecimal;
-
 public record SummaryPointResponse(
         String label,
-        BigDecimal total) {
+        double total) {
     public static SummaryPointResponse from(SummaryPoint point) {
         return new SummaryPointResponse(point.label(), point.total());
     }
